CC/CV Taper Function
This function tapers the charge current once the target voltage is
reached.

It will decrease current to hold voltage and subsequently increase
current to raise it, ideally holding the voltage at the target.

IF the minimumCharge current is ZERO, it will simply terminate the
charge when target voltage is reached.

The current increase function may be somewhat redundant with the rampup
in sendChademoStatus.

No change to  dynamic memory space.