CC/CV Taper Routines
This change adds routines to cut back current once the target voltage
is reached or increase current if we fall below the target voltage.

If minimumCurrent is set to ZERO, we will simply charge to target
voltage and terminate.

Part of this may be redundant with sendChademo Status.

No change on dynamic memory.