CAPACiTY and SOC
Added CAPACITY and SOC. You can define these up front in DEFINES but we
need them in the 404 message for the EVIC code in GEVCU.

Also, in addition to zeroing AH and kilowatt hours with the Z key, we
now also reset SOC to 100.

Finally, we added an autoreset.  EVERY time we reach the target charge
voltage, we reset AH to zero, kWh to zero, and SOC to 100%.  This is so
you don’t have to go in with a laptop to zero these values after every
charge.
In normal operation, your measured voltae should never reach your
target Charge voltage.  But at the end of charge it will.  So we
automatically reset these ongoing counters at the end of charge.

The nature of CC/CV is such that we will KEEP resetting each pass as we
continue charging and holding the constant voltage value.

But we will be at zero at charge completion this way.